+++ app/system/modules/settings/app/components/system.vue
@@ -25,6 +25,22 @@
     </div>
 
     <div class="uk-form-row">
+        <label for="form-user-recaptcha-enable" class="uk-form-label">{{ 'Google reCAPTCHA' | trans }}</label>
+        <div class="uk-form-controls uk-form-controls-text">
+            <p class="uk-form-controls-condensed">
+                <label><input id="form-user-recaptcha-enable" type="checkbox" v-model="$root.options['system/captcha'].recaptcha_enable"> {{ 'Enable for user registration and comments' | trans }}</label>
+            </p>
+            <p class="uk-form-controls-condensed" v-if="$root.options['system/captcha'].recaptcha_enable">
+                <input id="form-user-recaptcha-sitekey" class="uk-form-width-large" placeholder="{{ 'Site key' | trans }}" v-model="$root.options['system/captcha'].recaptcha_sitekey">
+            </p>
+            <p class="uk-form-controls-condensed" v-if="$root.options['system/captcha'].recaptcha_enable">
+                <input id="form-user-recaptcha-secret" class="uk-form-width-large" placeholder="{{ 'Secret key' | trans }}" v-model="$root.options['system/captcha'].recaptcha_secret">
+            </p>
+            <p class="uk-form-help-block">{{ 'Only key pairs for Google reRECAPTCHA V2 Invisible are supported.' | trans }}</p>
+        </div>
+    </div>
+
+    <div class="uk-form-row">
         <span class="uk-form-label">{{ 'Developer' | trans }}</span>
         <div class="uk-form-controls uk-form-controls-text">
             <p class="uk-form-controls-condensed">