+++ app/system/modules/user/src/Controller/RegistrationController.php
@@ -19,6 +19,9 @@ class RegistrationController
         $this->module = App::module('system/user');
     }
 
+    /**
+     * @Captcha(route="@user/registration/register")
+     */
     public function indexAction()
     {
         if (App::user()->isAuthenticated()) {
@@ -39,6 +42,7 @@ class RegistrationController
 
     /**
      * @Request({"user": "array"})
+     * @Captcha(verify="true")
      */
     public function registerAction($data)
     {